refactor(member, problem): service -> application 으로 패키지 네임 수정
