+++ src/main/java/com/lhgpds/algometa/configuration/security/handler/OAuth2SuccessHandler.java
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.lhgpds.algometa.controller.auth.dto.TokenDto;
 import com.lhgpds.algometa.exception.common.member.MemberDuplicateException;
 import com.lhgpds.algometa.internal.auth.jwt.service.JwtTokenService;
-import com.lhgpds.algometa.internal.member.service.MemberService;
-import com.lhgpds.algometa.internal.member.service.dto.MemberDto;
+import com.lhgpds.algometa.internal.member.application.MemberService;
+import com.lhgpds.algometa.internal.member.application.dto.MemberDto;
 import com.lhgpds.algometa.mapper.OAuth2Mapper;
 import java.util.Collections;
 import javax.servlet.http.Cookie;
