+++ src/main/java/com/lhgpds/algometa/controller/member/MemberController.java
@@ -8,8 +8,8 @@ import com.lhgpds.algometa.controller.member.dto.ResponseUploadImage;
 import com.lhgpds.algometa.infra.s3.S3Uploader;
 import com.lhgpds.algometa.internal.auth.jwt.principal.AlgoUser;
 import com.lhgpds.algometa.internal.member.domain.vo.ImageLink;
-import com.lhgpds.algometa.internal.member.service.MemberService;
-import com.lhgpds.algometa.internal.member.service.dto.MemberDto;
+import com.lhgpds.algometa.internal.member.application.MemberService;
+import com.lhgpds.algometa.internal.member.application.dto.MemberDto;
 import com.lhgpds.algometa.mapper.MemberMapper;
 import java.io.IOException;
 import lombok.RequiredArgsConstructor;
