+++ src/main/java/com/lhgpds/algometa/controller/problem/ProblemController.java
@@ -3,9 +3,9 @@ package com.lhgpds.algometa.controller.problem;
 import com.lhgpds.algometa.controller.problem.dto.RequestAddProblem;
 import com.lhgpds.algometa.controller.problem.dto.ResponseProblemId;
 import com.lhgpds.algometa.internal.auth.jwt.principal.AlgoUser;
-import com.lhgpds.algometa.internal.member.service.dto.MemberDto;
-import com.lhgpds.algometa.internal.problem.service.ProblemService;
-import com.lhgpds.algometa.internal.problem.service.dto.ProblemDto;
+import com.lhgpds.algometa.internal.member.application.dto.MemberDto;
+import com.lhgpds.algometa.internal.problem.application.ProblemService;
+import com.lhgpds.algometa.internal.problem.application.dto.ProblemDto;
 import com.lhgpds.algometa.mapper.ProblemMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
