+++ src/main/java/com/lhgpds/algometa/internal/auth/jwt/principal/AlgoUser.java
@@ -1,7 +1,7 @@
 package com.lhgpds.algometa.internal.auth.jwt.principal;
 
 
-import com.lhgpds.algometa.internal.member.service.dto.MemberDto;
+import com.lhgpds.algometa.internal.member.application.dto.MemberDto;
 import java.util.Collection;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.User;
