+++ src/main/java/com/lhgpds/algometa/internal/auth/jwt/service/JwtUserDetailsServiceImpl.java
@@ -4,7 +4,7 @@ import com.lhgpds.algometa.internal.auth.jwt.principal.AlgoUser;
 import com.lhgpds.algometa.internal.member.domain.entity.Member;
 import com.lhgpds.algometa.internal.member.domain.vo.Email;
 import com.lhgpds.algometa.internal.member.repository.MemberRepository;
-import com.lhgpds.algometa.internal.member.service.dto.MemberDto;
+import com.lhgpds.algometa.internal.member.application.dto.MemberDto;
 import com.lhgpds.algometa.mapper.MemberMapper;
 import java.util.Collections;
 import java.util.List;
