+++ src/main/java/com/lhgpds/algometa/internal/member/application/MemberServiceImpl.java
@@ -1,11 +1,11 @@
-package com.lhgpds.algometa.internal.member.service;
+package com.lhgpds.algometa.internal.member.application;
 
 import com.lhgpds.algometa.exception.common.member.MemberDuplicateException;
 import com.lhgpds.algometa.exception.common.member.MemberNotFoundException;
 import com.lhgpds.algometa.internal.member.domain.entity.Member;
 import com.lhgpds.algometa.internal.member.domain.vo.Role;
 import com.lhgpds.algometa.internal.member.repository.MemberRepository;
-import com.lhgpds.algometa.internal.member.service.dto.MemberDto;
+import com.lhgpds.algometa.internal.member.application.dto.MemberDto;
 import com.lhgpds.algometa.mapper.MemberMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
