+++ src/test/groovy/com/lhgpds/algometa/configuration/security/handler/OAuth2SuccessHandlerTest.groovy
@@ -6,8 +6,8 @@ import com.lhgpds.algometa.exception.common.member.MemberDuplicateException
 import com.lhgpds.algometa.internal.auth.jwt.service.JwtTokenService
 import com.lhgpds.algometa.internal.member.domain.vo.Email
 import com.lhgpds.algometa.internal.member.domain.vo.Role
-import com.lhgpds.algometa.internal.member.service.MemberServiceImpl
-import com.lhgpds.algometa.internal.member.service.dto.MemberDto
+import com.lhgpds.algometa.internal.member.application.MemberServiceImpl
+import com.lhgpds.algometa.internal.member.application.dto.MemberDto
 import org.springframework.http.MediaType
 import org.springframework.mock.web.MockHttpServletRequest
 import org.springframework.mock.web.MockHttpServletResponse
