+++ src/test/java/com/lhgpds/algometa/controller/member/MemberControllerDocsTest.java
@@ -25,8 +25,8 @@ import com.lhgpds.algometa.internal.member.domain.vo.Email;
 import com.lhgpds.algometa.internal.member.domain.vo.ImageLink;
 import com.lhgpds.algometa.internal.member.domain.vo.Nickname;
 import com.lhgpds.algometa.internal.member.domain.vo.Role;
-import com.lhgpds.algometa.internal.member.service.MemberService;
-import com.lhgpds.algometa.internal.member.service.dto.MemberDto;
+import com.lhgpds.algometa.internal.member.application.MemberService;
+import com.lhgpds.algometa.internal.member.application.dto.MemberDto;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
