+++ src/test/java/com/lhgpds/algometa/controller/problem/ProblemControllerDocsTest.java
@@ -25,8 +25,8 @@ import com.lhgpds.algometa.internal.problem.domain.vo.code.Difficulty;
 import com.lhgpds.algometa.internal.problem.domain.vo.code.Language;
 import com.lhgpds.algometa.internal.problem.domain.vo.code.Status;
 import com.lhgpds.algometa.internal.problem.domain.vo.content.Content;
-import com.lhgpds.algometa.internal.problem.service.ProblemService;
-import com.lhgpds.algometa.internal.problem.service.dto.ProblemDto;
+import com.lhgpds.algometa.internal.problem.application.ProblemService;
+import com.lhgpds.algometa.internal.problem.application.dto.ProblemDto;
 import java.time.LocalDateTime;
 import java.util.LinkedHashSet;
 import java.util.List;
